Pulling rule logic down into the extension
Before we were passing the target users down via the `child_token`. This posed a
few vectors for attack - the payload could be forged and linked to others. So
instead we are only sending down the current user via their token and then will
query the API again to find matching users.

I will likely sign and/or encrypt this token with the `EXTENSION_SECRET` or
something similar in the next iteration.

I also took this opportunity to refactor the code that loads up the template
into its own file under `template` with some separated requires in there to make
it easier to read at a glance.

There is now an acceptance test that verifies our extension fails properly with
a 400 if an invalid token was provided.

Overall this commit is a lot bigger than I wanted but it gets things done.
