+++ rules/link.js
@@ -5,7 +5,6 @@
  */
 export default ({extensionURL = '', username = 'Unknown', clientID = '', clientSecret = ''}) => {
   const template = `function (user, context, callback) {
-  console.log(LOG_TAG, 'Entered Account Link Rule');
   /**
    * This rule has been automatically generated by
    * ${username} at ${new Date().toISOString()}
@@ -16,6 +15,7 @@ export default ({extensionURL = '', username = 'Unknown', clientID = '', clientS
 
   var CONTINUE_PROTOCOL = 'redirect-callback';
   var LOG_TAG = '[ACCOUNT_LINK]: ';
+  console.log(LOG_TAG, 'Entered Account Link Rule');
 
   var config = {
     endpoints: {
@@ -36,6 +36,7 @@ export default ({extensionURL = '', username = 'Unknown', clientID = '', clientS
       return linkAccounts();
     } else if (shouldPrompt()) {
       return promptUser();
+
     }
 
     return continueAuth();
@@ -137,7 +138,7 @@ export default ({extensionURL = '', username = 'Unknown', clientID = '', clientS
     }).then(function redirectToExtension(targetUsers) {
       if (targetUsers.length > 0) {
         context.redirect = {
-          url: buildRedirectUrl(createToken(config.token, targetUsers), context.request.query)
+          url: buildRedirectUrl(createToken(config.token), context.request.query)
         };
       }
     });
@@ -165,7 +166,6 @@ export default ({extensionURL = '', username = 'Unknown', clientID = '', clientS
     var userSub = {
       sub: user.user_id,
       email: user.email,
-      targetUsers: targetUsers,
       base: auth0.baseUrl
     };
 
@@ -193,7 +193,7 @@ export default ({extensionURL = '', username = 'Unknown', clientID = '', clientS
       auth0Client: q.auth0Client,
       original_state: q.original_state || q.state,
       nonce: q.nonce,
-      errorType: errorType
+      error_type: errorType
     };
 
     return config.endpoints.linking + '?' + queryString.encode(params);
