--- templates/layout/layout.pug
@@ -1,9 +1,0 @@
-html
-  head
-    title My Site - #{title}
-    block scripts
-  body
-    block content
-    block foot
-      #footer
-        p some footer content
