--- templates/link.pug
@@ -1,5 +1,0 @@
-extends layout.pug
-
-block content
-  h1 Hello!
-  p= message
